requirements.txt
python-telegram-bot==12.8
python-decouple==3.4